replace FieldSpecificationBase with FieldSpecification
ProppantTransport.hpp and PhaseFieldDamageFEM.hpp both had the "class FieldSpecification;" line that would cause an error with the temporary alias in FieldSpecification.hpp. That's why they are in this commit.
