+++ src/coreComponents/fieldSpecification/FieldSpecification.hpp
@@ -21,11 +21,348 @@
 #define GEOS_FIELDSPECIFICATION_FIELDSPECIFICATION_HPP
 
 
-#include "FieldSpecificationBase.hpp"
+#include "common/DataTypes.hpp"
+#include "common/TypeDispatch.hpp"
+#include "codingUtilities/traits.hpp"
+#include "codingUtilities/Utilities.hpp"
+#include "dataRepository/Group.hpp"
+#include "functions/FunctionBase.hpp"
+#include "common/FieldSpecificationOps.hpp"
+#include "mesh/ObjectManagerBase.hpp"
+#include "mesh/MeshObjectPath.hpp"
+#include "functions/FunctionManager.hpp"
+#include "common/GEOS_RAJA_Interface.hpp"
 
-namespace geos 
+namespace geos
 {
-using FieldSpecification = geos::FieldSpecificationBase;
+class Function;
+
+
+/**
+ * @class FieldSpecification
+ * A class to hold values for and administer a single boundary condition
+ */
+class FieldSpecification : public dataRepository::Group
+{
+public:
+
+  /**
+   * @defgroup alias and functions to defined statically initialized catalog
+   * @{
+   */
+
+  /**
+   * alias to define the catalog type for this base type
+   */
+  using CatalogInterface = dataRepository::CatalogInterface< FieldSpecification,
+                                                             string const &,
+                                                             dataRepository::Group * const >;
+  /**
+   * @enum  SetErrorMode
+   * @brief Indicate the error handling mode.
+   */
+  enum class SetErrorMode : integer
+  {
+    silent,
+    error,
+    warning
+  };
+
+  /**
+   * @brief static function to return static catalog.
+   * @return the static catalog to create derived types through the static factory methods.
+   */
+  static CatalogInterface::CatalogType & getCatalog();
+
+  /**
+   * @brief Static Factory Catalog Functions
+   * @return the catalog name
+   */
+  static string catalogName() { return "FieldSpecification"; }
+
+  /**
+   * @brief return the catalog name
+   * @return the catalog name
+   */
+  virtual const string getCatalogName() const
+  {
+    return FieldSpecification::catalogName();
+  }
+
+  /**
+   * @}
+   */
+
+
+  /**
+   * @brief constructor
+   * @param name the name of the FieldSpecification in the data repository
+   * @param parent the parent group of this group.
+   */
+  FieldSpecification( string const & name, dataRepository::Group * parent );
+
+  /**
+   * destructor
+   */
+  virtual ~FieldSpecification() override;
+
+
+  /// Deleted copy constructor
+  FieldSpecification( FieldSpecification const & ) = delete;
+
+  /// Defaulted move constructor
+  FieldSpecification( FieldSpecification && ) = default;
+
+  /// deleted copy assignment
+  FieldSpecification & operator=( FieldSpecification const & ) = delete;
+
+  /// deleted move assignement
+  FieldSpecification & operator=( FieldSpecification && ) = delete;
+
+  /**
+   * @brief View keys
+   */
+  struct viewKeyStruct
+  {
+    /// @return The key for setName
+    constexpr static char const * setNamesString() { return "setNames"; }
+    /// @return The key for constitutivePath
+    constexpr static char const * constitutivePathString() { return "constitutivePath"; }
+    /// @return The key for objectPath
+    constexpr static char const * objectPathString() { return "objectPath"; }
+    /// @return The key for fieldName
+    constexpr static char const * fieldNameString() { return "fieldName"; }
+    /// @return The key for dataType
+    constexpr static char const * dataTypeString() { return "dataType"; }
+    /// @return The key for component
+    constexpr static char const * componentString() { return "component"; }
+    /// @return The key for direction
+    constexpr static char const * directionString() { return "direction"; }
+    /// @return The key for bcApplicationTableName
+    constexpr static char const * bcApplicationTableNameString() { return "bcApplicationTableName"; }
+    /// @return The key for scale
+    constexpr static char const * scaleString() { return "scale"; }
+    /// @return The key for functionName
+    constexpr static char const * functionNameString() { return "functionName"; }
+    /// @return The key for initialCondition
+    constexpr static char const * initialConditionString() { return "initialCondition"; }
+    /// @return The key for beginTime
+    constexpr static char const * beginTimeString() { return "beginTime"; }
+    /// @return The key for endTime
+    constexpr static char const * endTimeString() { return "endTime"; }
+    /// @return The key errorSetMode
+    constexpr static char const * errorSetModeString() { return "errorSetMode"; }
+  };
+
+  /**
+   * Accessor
+   * @return const reference to m_function
+   */
+  string const & getFunctionName() const
+  {
+    return m_functionName;
+  }
+
+  /**
+   * Accessor
+   * @return const reference to m_objectPath
+   */
+  virtual const string & getObjectPath() const
+  {
+    return m_objectPath;
+  }
+
+  /**
+   * Accessor
+   * @return const reference to m_fieldName
+   */
+  virtual const string & getFieldName() const
+  {
+    return m_fieldName;
+  }
+
+  /**
+   * Accessing the considered component.
+   * @return The component axis or a special value.
+   */
+  virtual int getComponent() const
+  {
+    return m_component;
+  }
+
+  /**
+   * Accessor
+   * @return const reference to m_direction
+   */
+  virtual R1Tensor const & getDirection() const
+  {
+    GEOS_UNUSED_VAR( time );
+    return m_direction;
+  }
+
+  /**
+   * Accessor
+   * @return const m_beginTime
+   */
+  real64 getStartTime() const
+  {
+    return m_beginTime;
+  }
+
+  /**
+   * Accessor
+   * @return const m_endTime
+   */
+  real64 getEndTime() const
+  {
+    return m_endTime;
+  }
+
+  /**
+   * Accessor
+   * @return const reference to m_setNames
+   */
+  string_array const & getSetNames() const
+  {
+    return m_setNames;
+  }
+
+  /**
+   * Accessor
+   * @return const m_initialCondition
+   */
+  int initialCondition() const
+  {
+    return m_initialCondition;
+  }
+
+  /**
+   * Accessor
+   * @return const m_scale
+   */
+  real64 getScale() const
+  {
+    return m_scale;
+  }
+
+  /**
+   * Mutator
+   * @param[in] fieldName The name of the field
+   */
+  void setFieldName( string const & fieldName )
+  {
+    m_fieldName = fieldName;
+  }
+
+  /**
+   * Mutator
+   * @param[in] objectPath The path for the object
+   */
+  void setObjectPath( string const & objectPath )
+  {
+    m_objectPath = objectPath;
+  }
+
+  /**
+   * Mutator
+   * @param[in] scale Scaling factor
+   */
+  void setScale( real64 const & scale )
+  {
+    m_scale = scale;
+  }
+
+  /**
+   * Mutator
+   * @param[in] isInitialCondition Logical value to indicate if it is an initial condition
+   */
+  void initialCondition( bool isInitialCondition )
+  {
+    m_initialCondition = isInitialCondition;
+  }
+
+  /**
+   * Mutator
+   * @param[in] setName The name of the set
+   */
+  void addSetName( string const & setName )
+  {
+    m_setNames.emplace_back( setName );
+  }
+
+  /**
+   * @brief Set the Mesh Object Path object
+   *
+   * @param meshBodies The group containing all the MeshBody objects
+   */
+  void setMeshObjectPath( Group const & meshBodies );
+
+  /**
+   * @brief Get the Mesh Object Paths object
+   *
+   * @return reference to const m_meshObjectPaths
+   */
+  MeshObjectPath const & getMeshObjectPaths() const
+  {
+    return *(m_meshObjectPaths.get());
+  }
+
+
+protected:
+
+
+private:
+
+
+  /// the names of the sets that the boundary condition is applied to
+  string_array m_setNames;
+
+  /// the path to the object which contains the fields that the boundary condition is applied to
+  string m_objectPath;
+
+  std::unique_ptr< MeshObjectPath > m_meshObjectPaths;
+
+  /// the name of the field the boundary condition is applied to or a key string to use for
+  /// determining whether or not to apply the boundary condition.
+  string m_fieldName;
+
+
+  /// The component the boundary condition acts on. Not used if field is a scalar.
+  int m_component;
+
+  /// The direction the boundary condition acts in.
+  R1Tensor m_direction;
+
+  /// Whether or not the boundary condition is an initial condition.
+  int m_initialCondition;
+
+  /// The name of the function used to generate values for application.
+  string m_functionName;
+
+  /// The scale factor to use on the value of the boundary condition.
+  real64 m_scale;
+
+  /// Time after which the bc is allowed to be applied
+  real64 m_beginTime;
+
+  /// Time after which the bc will no longer be applied.
+  real64 m_endTime;
+
+  /// The name of a function used to turn on and off the boundary condition.
+  string m_bcApplicationFunctionName;
+
+  /// Enum containing the possible output modes when an error occur
+  SetErrorMode m_emptySetErrorMode;
+};
+
+/**
+ * @brief Indicate the error handling mode
+ */
+ENUM_STRINGS( FieldSpecification::SetErrorMode,
+              "silent",
+              "error",
+              "warning" );
+
 }
 
 #endif //GEOS_FIELDSPECIFICATION_FIELDSPECIFICATION_HPP
