+++ src/coreComponents/physicsSolvers/simplePDE/PhaseFieldDamageFEM.hpp
@@ -36,7 +36,7 @@ namespace dataRepository
 {
 class Group;
 }
-class FieldSpecificationBase; // TODO: change to FieldSpecification
+class FieldSpecification;
 class FiniteElementBase;
 class DomainPartition;
